requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.7
tiktoken==0.5.2
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
